+++ src/realm/alloc_slab.cpp
@@ -64,6 +64,7 @@ struct SlabAlloc::MappedFile {
 
     util::Mutex m_mutex;
     util::File m_file;
+    util::SharedFileInfo* m_realm_file_info = nullptr;
     util::File::Map<char> m_initial_mapping;
     // additional sections beyond those covered by the initial mapping, are
     // managed as separate mmap allocations, each covering one section.
@@ -1076,9 +1077,23 @@ ref_type SlabAlloc::attach_file(const std::string& file_path, Config& cfg)
     }
     dg.release();  // Do not detach
     fcg.release(); // Do not close
+    m_file_mappings->m_realm_file_info = util::get_file_info_for_file(m_file_mappings->m_file);
     m_file_mappings->m_success = true;
     return top_ref;
 }
+
+void SlabAlloc::note_reader_start(void* reader_id)
+{
+    if (m_file_mappings->m_realm_file_info)
+        util::encryption_note_reader_start(*m_file_mappings->m_realm_file_info, reader_id);
+}
+
+void SlabAlloc::note_reader_end(void* reader_id)
+{
+    if (m_file_mappings->m_realm_file_info)
+        util::encryption_note_reader_end(*m_file_mappings->m_realm_file_info, reader_id);
+}
+
 
 ref_type SlabAlloc::attach_buffer(const char* data, size_t size)
 {