+++ src/realm/group_writer.cpp
@@ -129,6 +129,7 @@ GroupWriter::MapWindow::MapWindow(size_t alignment, util::File& f, ref_type star
 
 GroupWriter::MapWindow::~MapWindow()
 {
+    m_map.unmap(); /* Apparently no effect - how odd */
 }
 
 void GroupWriter::MapWindow::sync()